+++ mobile/lib/features/fortune/presentation/pages/fortune_tarot_hub_page.dart
@@ -1,3 +1,4 @@
+import 'package:canlifal_social/core/performance/animation_perf.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
@@ -25,26 +26,18 @@ class FortuneTarotHubPage extends ConsumerStatefulWidget {
 
 class _FortuneTarotHubPageState extends ConsumerState<FortuneTarotHubPage> {
   final _scrollController = ScrollController();
-  double _scrollOffset = 0;
+  final _scrollParallax = ScrollParallaxNotifier();
 
   @override
   void initState() {
     super.initState();
-    _scrollController.addListener(_onScroll);
-  }
-
-  void _onScroll() {
-    final offset = _scrollController.offset;
-    if ((offset - _scrollOffset).abs() > 0.5) {
-      setState(() => _scrollOffset = offset);
-    }
+    _scrollParallax.bind(_scrollController);
   }
 
   @override
   void dispose() {
-    _scrollController
-      ..removeListener(_onScroll)
-      ..dispose();
+    _scrollParallax.dispose();
+    _scrollController.dispose();
     super.dispose();
   }
 
@@ -63,7 +56,7 @@ class _FortuneTarotHubPageState extends ConsumerState<FortuneTarotHubPage> {
     return Scaffold(
       backgroundColor: bg,
       body: UltraFortuneCosmicBackground(
-        scrollOffset: _scrollOffset,
+        scrollParallax: _scrollParallax,
         child: DiscoverRefresh.wrap(
           onRefresh: _onRefresh,
           child: CustomScrollView(